backend/requirements.txt
pytest>=7.0
pytest-cov>=4.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0
httpx>=0.24.0
PyJWT>=2.8.0
cryptography>=41.0.0
python-json-logger>=2.0.0
